// src/card_request.interface.ts
export interface CardRequest {
    readonly name: string;
    readonly email: string;
    readonly phone: string;
    readonly cpf: string;
    readonly urgency: string;
    readonly contactTime: string;
    readonly wantToSpend: string;
    readonly money: string;
    readonly typeCard: string;
  }
  